Fixed channel-wise data getter.
When optional argument `channels` is provided, there's the possibility
that one of the requested channel labels is in non of the real labels,
but that it has been added to the `channel_by_label` dict.  Then,
the returned dict with the data should still contain the label in
`channels`. (and not the original one;
`channel_by_label[channels[i]].label`)